requirements.txt
hydra-core==1.3.2
PyWavefront==1.3.3
PyYAML==6.0
taichi==1.6.0
tqdm==4.65.0
pin==2.6.18
cmeel-boost==1.82.0
sapien==2.2.1
zarr==2.14.2
torch==2.0.0
lightning==2.0.3
tensorboard==2.13.0
transformers==4.30.2
matplotlib==3.7.4  